fix mpfr
